Provide default download time
If no matching FW model is found, provide the maximum FW size
for client to have valid estimated download time instead of error.

Also, we are inserting the missing FW revision to DB with size 0
which would serve as our flag to replace/populate it with correct size later on.